+++ native/xqlitenif/src/atoms.rs
@@ -1,0 +1,13 @@
+rustler::atoms! {
+    already_closed,
+    cannot_close,
+    cannot_decode,
+    cannot_execute,
+    db_name,
+    error,
+    nil,
+    ok,
+    pragma_get_failed,
+    pragma_put_failed,
+    unsupported_pragma_put_value
+}
