+++ native/xqlitenif/src/close.rs
@@ -1,0 +1,43 @@
+use rusqlite::Connection;
+use rustler::resource::ResourceArc;
+use rustler::{Encoder, Env, Term};
+
+use crate::atoms::{already_closed, cannot_close, error, ok};
+use crate::shared::XqliteConnection;
+
+enum CloseResult {
+    Success,
+    AlreadyClosed,
+    Failure(String),
+}
+
+impl<'a> Encoder for CloseResult {
+    fn encode<'b>(&self, env: Env<'b>) -> Term<'b> {
+        match self {
+            CloseResult::Success => (ok()).encode(env),
+            CloseResult::AlreadyClosed => (error(), already_closed()).encode(env),
+            CloseResult::Failure(msg) => (error(), cannot_close(), msg).encode(env),
+        }
+    }
+}
+
+#[rustler::nif(schedule = "DirtyIo")]
+fn close(arc: ResourceArc<XqliteConnection>) -> CloseResult {
+    let mut mconn = arc.0.lock().unwrap();
+    if !mconn.is_none() {
+        // Take out the connection so it becomes None in the resource
+        // as it is being closed early.
+        let conn: Connection = mconn.take().unwrap();
+        match conn.close() {
+            Ok(()) => CloseResult::Success,
+            Err((conn, err)) => {
+                // Closing failed, put the connection back in the original container.
+                *mconn = Some(conn);
+                let msg: String = format!("{:?}", err);
+                CloseResult::Failure(msg)
+            }
+        }
+    } else {
+        CloseResult::AlreadyClosed
+    }
+}
